+++ src/layouting/layouting.jl
@@ -8,7 +8,9 @@ end
 
 function attribute_per_char(string, attribute)
     n_words = 0
-    if attribute isa AbstractVector
+    if attribute isa GeometryBasics.StaticArray
+        return one_attribute_per_char(attribute, string)
+    elseif attribute isa AbstractVector
         if length(attribute) == length(string)
             return attribute
         else
@@ -83,7 +85,7 @@ function glyph_collection(
             char = char,
             font = font,
             scale = scale,
-            lineheight = Float32(font.height / font.units_per_EM * lineheight_factor * scale),
+            lineheight = Float32(font.height / font.units_per_EM * lineheight_factor * last(scale)),
             extent = GlyphExtent(font, char)
         )
     end
@@ -111,7 +113,7 @@ function glyph_collection(
 
         for (i, ci) in enumerate(charinfos)
             push!(xs[end], x)
-            x += ci.extent.hadvance * ci.scale
+            x += ci.extent.hadvance * first(ci.scale)
 
             if 0 < word_wrap_width < x && last_space_local_idx != 0 &&
                     ((ci.char in (' ', '\n')) || i == length(charinfos))
@@ -121,7 +123,7 @@ function glyph_collection(
                 xs[end-1] = xs[end-1][1:last_space_local_idx]
                 push!(lineinfos, view(charinfos, last_line_start:last_space_global_idx))
                 last_line_start = last_space_global_idx+1
-                x = xs[end][end] + ci.extent.hadvance * ci.scale
+                x = xs[end][end] + ci.extent.hadvance * first(ci.scale)
 
                 # TODO Do we need to redo the metrics for newlines?
                 charinfos[last_space_global_idx] = let
@@ -156,7 +158,7 @@ function glyph_collection(
         # if the last and not the only character is \n, take the previous one
         # to compute the width
         i = (nchars > 1 && line[end].char == '\n') ? nchars - 1 : nchars
-        xx[i] + line[i].extent.hadvance * line[i].scale
+        xx[i] + line[i].extent.hadvance * first(line[i].scale)
     end
 
     # the maximum width is needed for justification
@@ -217,11 +219,11 @@ function glyph_collection(
     # for y alignment, we need the largest ascender of the first line
     # and the largest descender of the last line
     first_line_ascender = maximum(lineinfos[1]) do l
-        l.scale * l.extent.ascender
+        last(l.scale) * l.extent.ascender
     end
 
     last_line_descender = minimum(lineinfos[end]) do l
-        l.scale * l.extent.descender
+        last(l.scale) * l.extent.descender
     end
 
     # compute the height of all lines together