+++ ReferenceTests/src/tests/examples2d.jl
@@ -200,7 +200,7 @@ end
     sf = Observable(Base.Fix2(v, 0e0))
     title_str = Observable("t = 0.00")
     sp = streamplot(sf, -2..2, -2..2;
-                    linewidth=2,  arrow_size=20, colormap=:magma, axis=(;title=title_str))
+                    linewidth=2, colormap=:magma, axis=(;title=title_str))
     Record(sp, LinRange(0, 20, 5)) do i
         sf[] = Base.Fix2(v, i)
         title_str[] = "t = $(round(i; sigdigits=2))"
@@ -268,6 +268,63 @@ end
     end
 
     Makie.step!(st)
+    st
+end
+
+@reference_test "Axes label rotations" begin
+    axis = (
+        xlabel = "a long x label for this axis",
+        ylabel = "a long y\nlabel for this axis",
+        xlabelrotation = π / 4,
+        ylabelrotation = 0,
+    )
+    fig, ax, _ = scatter(0:1; axis)
+
+    st = Stepper(fig)
+    Makie.step!(st)
+
+    ax.yaxisposition[] = :right
+    ax.ylabelrotation[] = Makie.automatic
+    ax.xlabelrotation[] = -π / 5
+    Makie.step!(st)
+
+    ax.xaxisposition[] = :top
+    ax.xlabelrotation[] = 3π / 4
+    ax.ylabelrotation[] = π / 4
+    Makie.step!(st)
+
+    # reset to defaults
+    ax.xaxisposition[] = :bottom
+    ax.yaxisposition[] = :left
+    ax.xlabelrotation[] = ax.ylabelrotation[] = Makie.automatic
+    Makie.step!(st)
+
+    st
+end
+
+@reference_test "Colorbar label rotations" begin
+    axis = (
+        xlabel = "x axis label",
+        ylabel = "y axis label",
+        xlabelrotation = -π / 10,
+        ylabelrotation = -π / 3,
+        yaxisposition = :right,
+    )
+    fig, _, _ = scatter(0:1; axis)
+
+    cb_vert = Colorbar(fig[1, 2]; label = "vertical cbar", labelrotation = 0)
+    cb_horz = Colorbar(fig[2, 1]; label = "horizontal cbar", labelrotation = π / 5, vertical = false)
+
+    st = Stepper(fig)
+    Makie.step!(st)
+
+    # reset to defaults
+    cb_vert.labelrotation[] = Makie.automatic
+    Makie.step!(st)
+
+    cb_horz.labelrotation[] = Makie.automatic
+    Makie.step!(st)
+
     st
 end
 
@@ -557,7 +614,7 @@ end
     )
     tooltip!(ax, Point2f(0), "below", placement = :below, outline_color = :red, outline_linestyle = :dot)
     tooltip!(
-        ax, 0, 0, text = "right", placement = :right, fontsize = 30, 
+        ax, 0, 0, text = "right", placement = :right, fontsize = 30,
         outline_linewidth = 5, offset = 30, triangle_size = 15,
         strokewidth = 2f0, strokecolor = :cyan
     )