+++ NEWS.md
@@ -3,6 +3,7 @@
 ## master
 
 - Added `xlabelrotation`, `ylabelrotation` (`Axis`) and `labelrotation` (`Colorbar`) [#2478](https://github.com/MakieOrg/Makie.jl/pull/2478).
+- Fixed forced rasterization in CairoMakie svg files when polygons with colors specified as (color, alpha) tuples were used [#2535](https://github.com/MakieOrg/Makie.jl/pull/2535).
 - Added support for `linecap` to CairoMakie and GLMakie [#2536](https://github.com/MakieOrg/Makie.jl/pull/2536)
 - Added support for per-position `linewidth` to `lines!` in GLMakie [#2536](https://github.com/MakieOrg/Makie.jl/pull/2536)